refactor: split files of widget helpers
